backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
boto3
pydantic
pypdf
python-multipart
requests
